Door coordinates are now random.
There will need to be a deep discussion on how the doors will function relative to player as things are created on the account that this is chaos, when we want controlled chaos. We need to know what doors are connected to rooms, and so on, so that this functions perfectly as intended.